Fix Undefined environments not being handled correctly. Refactored for easier Unit Testing
- Refactored for easier Unit Testing
- Added some simple Unit Tests
- Make sure that unset environments do not throw exceptions because exceptionlogger doesn't handle them
