+++ Notifier/Notifier.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Dopiaza\Slack\ExceptionLoggerBundle\Notifier;
+
+interface Notifier
+{
+    /**
+     * @param \Exception $exception
+     */
+    public function notify(\Exception $exception);
+}
