+++ composer.json
@@ -1,32 +1,31 @@
 {
-  "name" : "dopiaza/slack-exception-logger-bundle",
-  "description" : "Log symfony2 exceptions to a Slack channel",
-  "type" : "symfony-bundle",
-  "authors" : [{
-    "name" : "David Wilkinson",
-    "email" : "dopiaza@dopiaza.org"
-  }],
-  "keywords" : [
-    "exception", "slack"
+  "name": "dopiaza/slack-exception-logger-bundle",
+  "description": "Log symfony2 exceptions to a Slack channel",
+  "type": "symfony-bundle",
+  "authors": [
+    {
+      "name": "David Wilkinson",
+      "email": "dopiaza@dopiaza.org"
+    }
   ],
-  "license" : [
+  "keywords": [
+    "exception",
+    "slack"
+  ],
+  "license": [
     "MIT"
   ],
-  "require" : {
-      "php": ">=5.3.2",
-      "symfony/framework-bundle": "2.*"
+  "require": {
+    "php": ">=5.3.2",
+    "symfony/framework-bundle": "2.*"
   },
-  "autoload" : {
-    "psr-0" : {
-      "Dopiaza\\Slack\\ExceptionLoggerBundle" : ""
-    }
+  "require-dev": {
+    "phpunit/phpunit": "~4.8",
+    "mockery/mockery": "~0.9"
   },
-  "target-dir" : "Dopiaza/Slack/ExceptionLoggerBundle",
-  "repositories" : [{
-  }],
-  "extra" : {
-    "branch-alias" : {
-      "dev-master" : "1.0.x-dev"
+  "autoload": {
+    "psr-4": {
+      "Dopiaza\\Slack\\ExceptionLoggerBundle\\": ""
     }
   }
 }
